Spam and Phishing modules v2
- Restored extract_domain function
- Fixed comments within both programs
- Removed keyword "free" within the spam_keywords text file, as this keywords brings false positives.
- Changed implementation of is_spam on spamtest.py. Instead of iterating through the entire message, implemented it so that it instantly returns the result if it detects a keyword.
- Added error checking with the load_phishing_domains and load_phishing_links function
- Removed unused imports from both files
- Moved files into two separate folders based on usage: spam filter and phishing filter.
