+++ Modules/Spam Filter/spam_keywords.txt
@@ -1,7 +1,6 @@
 win
 winner
 prize
-free
 lottery
 guaranteed
 money back
